(1248) Split up permissionModel
