+++ packages/api/server/models/permissionModel/find.js
@@ -1,0 +1,10 @@
+const query = require('./_common/query');
+
+/**
+ * Returns a permission map for the requested owners only
+ *
+ * @param {PermissionOwners} owners
+ *
+ * @returns {PermissionMap}
+ */
+module.exports = owners => query(owners);
