+++ packages/api/server/models/permissionModel/findAll.js
@@ -1,0 +1,8 @@
+const query = require('./_common/query');
+
+/**
+ * Returns a full permission map
+ *
+ * @returns {PermissionMap}
+ */
+module.exports = query();
